Refactor to include AWS services in own package
