+++ formstack-baton-requests/src/local/com/gu/identity/formstackbatonrequests/FormstackBatonLambdaLocalRun.scala
@@ -5,13 +5,12 @@ import java.io.{ByteArrayInputStream, ByteArrayOutputStream}
 import BatonModels.{SarRequest, SarStatusRequest}
 import io.circe.syntax._
 import circeCodecs._
-import com.gu.identity.formstackbatonrequests.lambda.Lambda
-import com.gu.identity.formstackbatonrequests.s3.S3
+import com.gu.identity.formstackbatonrequests.aws.{Lambda, S3}
 
 /** This object can be used for local runs of the lambda, for end-to-end testing. */
 
 object FormstackBatonLambdaLocalRun extends App {
-  def runWith(request: SarRequest) = {
+  def runWith(request: SarRequest): Unit = {
     val sarLambdaConfig = FormstackConfig.getSarHandlerConfig
     val sarLambda = FormstackSarHandler(S3, Lambda, sarLambdaConfig)
     val jsonRequest = request.asJson.noSpaces
