+++ formstack-baton-requests/src/main/scala/com/gu/identity/formstackbatonrequests/aws/AwsCredentials.scala
@@ -1,0 +1,16 @@
+package com.gu.identity.formstackbatonrequests.aws
+
+import com.amazonaws.auth.profile.ProfileCredentialsProvider
+import com.amazonaws.auth.{AWSCredentialsProviderChain, EnvironmentVariableCredentialsProvider, InstanceProfileCredentialsProvider, SystemPropertiesCredentialsProvider}
+import com.amazonaws.regions.Regions
+
+object AwsCredentials {
+  def credentials = new AWSCredentialsProviderChain(
+    new EnvironmentVariableCredentialsProvider(),
+    new SystemPropertiesCredentialsProvider(),
+    new ProfileCredentialsProvider("identity"),
+    new InstanceProfileCredentialsProvider(false)
+  )
+
+  val region: Regions = Regions.EU_WEST_1
+}
